+++ CHANGELOG.md
@@ -9,9 +9,17 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Semantic Versioning]: https://semver.org/
 
 
-## [Unreleased](https://github.com/bbugyi200/clap/compare/0.4.0...HEAD)
+## [Unreleased](https://github.com/bbugyi200/clap/compare/0.5.0...HEAD)
 
 No notable changes have been made.
+
+
+## [0.5.0](https://github.com/bbugyi200/clap/compare/0.4.0...0.5.0) - 2021-12-21
+
+### Changed
+
+* Make `Config` attributes immutable.
+* *BREAKING CHANGE*: The `clap.main_factory(run, config)` function's parameters have changed.
 
 
 ## [0.4.0](https://github.com/bbugyi200/clap/compare/0.3.1...0.4.0) - 2021-12-20